% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{addR}
\alias{addR}
\title{Add a row to an array}
\usage{
addR(A, x)
}
\arguments{
\item{A}{Array}

\item{x}{Vector to add}
}
\value{
Numeric Matrix with A and added x as row
}
\description{
Given an mxn array and a vector, if size of vector equal to number of column of array, 
added the vector in to array as row
}
\examples{
myArray <- matrix(1:12, nrow = 3, ncol = 4);
auxV <- c(13,14,15,16);
NewArray <- addR(myArray, auxV);
NewArray;
}
