% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MaxMin}
\alias{MaxMin}
\title{Maxmin Convolution between two matrices equal squares}
\usage{
MaxMin(A, B)
}
\arguments{
\item{A}{Array}

\item{B}{Array}
}
\value{
Numeric Matrix with the maxmin elements
}
\description{
Get the minimum elements between the row of A with the column of B, 
then take the maximum value between the minimums and store it in an array C
}
\examples{
mA <- matrix(1:16, nrow = 4, ncol = 4);
mB <- matrix(16:1, nrow = 4, ncol = 4);
MaxMin(mA,mB);
}
