% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SearchMax}
\alias{SearchMax}
\title{Search of maximum Values}
\usage{
SearchMax(x)
}
\arguments{
\item{A}{Array}
}
\value{
Data frame with the index the maximum value
}
\description{
Given an nxn array, returns the maximum value found and the index where they are found
}
\examples{
MaxValue <- SearchMax(myMatrix);
}
