% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SearchMin}
\alias{SearchMin}
\title{Search of Minimum Values}
\usage{
SearchMin(x)
}
\arguments{
\item{A}{Array}
}
\value{
Data frame with the index the minimum value
}
\description{
Given an nxn array, returns the minimum value found and the index where they are found
}
\examples{
mA <- matrix(1:12, nrow = 3, ncol = 4);
SearchMin(mA);
}
