% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{addC}
\alias{addC}
\title{Add a column to an array}
\usage{
addC(A, x)
}
\arguments{
\item{A}{Array}

\item{x}{Vector to add}
}
\value{
Numeric Matrix with A and added x as column
}
\description{
Given an mxn array and a vector, if size of vector equal to number of row of array, 
added the vector in to array as column
}
\examples{
NewArray <- addColumn(myArray, c(1,2,3,4));
}
