% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MxM}
\alias{MxM}
\title{Matrix Multiplications}
\usage{
MxM(A, B)
}
\arguments{
\item{A}{Array}

\item{B}{Array}
}
\value{
Numeric Matrix with the result of A x B
}
\description{
Given a matrix A and a matrix B, returns the multiplication between A and B
}
\examples{
C <- MxM(myMatrix1 , myMatrix2);
}
